perf(home): convert home page to server component, split client island (#769)
Addresses #769 from epic #762.

HomePageClient was 435 LOC of 'use client' markup. Beyond the session-check
for the "Entra según tu perfil" block, nothing on the home page actually
needed to run on the client — but because the root was a client component,
every child (ProductCard, icons, translation hooks) was pulled into the
client bundle.

## Change

- page.tsx is now a server component that renders the whole home using
  getServerT() + getServerLocale(). Translations happen once on the server;
  the client gets static HTML for everything except the one interactive bit.
- New: src/components/home/QuickAccessIsland.tsx — a tiny 'use client'
  wrapper that reads useSession() and hides its children when the user
  is authenticated. Server still renders the block so anonymous visitors
  see a complete landing with no hydration flicker.
- HomePageClient.tsx deleted.

## Tradeoff

Reading cookies via getServerLocale pushes the page into dynamic rendering
(the old `export const revalidate = 3600` is gone, replaced by
`force-dynamic`). Data access still goes through getHomeSnapshot, which
wraps its catalog queries in unstable_cache(3600) — so the DB is untouched
on a warm cache. The only cost is per-request HTML generation, which is
fast; the payoff is removing the 435-LOC client bundle from the most-hit
page on the marketplace.

If the CDN-caching regression bites (e.g. Vercel edge cache loss), we
can move locale into a [locale] URL segment and restore ISR per locale.

typecheck + lint clean.
